Move conditional to the app, rather than component
